@@ -1,6 +1,6 @@
 name: example
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 dependencies:
   flutter_chat_types:
     path: ..
