@@ -2,19 +2,19 @@ name: flutter_chat_types
 description: >
   Utility library for the flutter_chat_ui and flutter_firebase_chat_core libraries
   which contains shared type declarations.
-version: 3.1.4
+version: 3.2.0
 homepage: https://flyer.chat
 repository: https://github.com/flyerhq/flutter_chat_types
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   equatable: ^2.0.3
-  json_annotation: ^4.0.0
+  json_annotation: ^4.3.0
   meta: ^1.3.0
 
 dev_dependencies:
-  build_runner: ^2.1.2
+  build_runner: ^2.1.4
   flutter_lints: ^1.0.4
-  json_serializable: ^5.0.0
+  json_serializable: ^6.0.1
