@@ -16,8 +16,8 @@ FileMessage _$FileMessageFromJson(Map<String, dynamic> json) => FileMessage(
       remoteId: json['remoteId'] as String?,
       roomId: json['roomId'] as String?,
       size: json['size'] as num,
-      status: _$enumDecodeNullable(_$StatusEnumMap, json['status']),
-      type: _$enumDecodeNullable(_$MessageTypeEnumMap, json['type']),
+      status: $enumDecodeNullable(_$StatusEnumMap, json['status']),
+      type: $enumDecodeNullable(_$MessageTypeEnumMap, json['type']),
       updatedAt: json['updatedAt'] as int?,
       uri: json['uri'] as String,
     );
@@ -38,43 +38,6 @@ Map<String, dynamic> _$FileMessageToJson(FileMessage instance) =>
       'size': instance.size,
       'uri': instance.uri,
     };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
-
-K? _$enumDecodeNullable<K, V>(
-  Map<K, V> enumValues,
-  dynamic source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
-}
 
 const _$StatusEnumMap = {
   Status.delivered: 'delivered',
