@@ -16,9 +16,9 @@ TextMessage _$TextMessageFromJson(Map<String, dynamic> json) => TextMessage(
           : PreviewData.fromJson(json['previewData'] as Map<String, dynamic>),
       remoteId: json['remoteId'] as String?,
       roomId: json['roomId'] as String?,
-      status: _$enumDecodeNullable(_$StatusEnumMap, json['status']),
+      status: $enumDecodeNullable(_$StatusEnumMap, json['status']),
       text: json['text'] as String,
-      type: _$enumDecodeNullable(_$MessageTypeEnumMap, json['type']),
+      type: $enumDecodeNullable(_$MessageTypeEnumMap, json['type']),
       updatedAt: json['updatedAt'] as int?,
     );
 
@@ -36,43 +36,6 @@ Map<String, dynamic> _$TextMessageToJson(TextMessage instance) =>
       'previewData': instance.previewData?.toJson(),
       'text': instance.text,
     };
-
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
-
-K? _$enumDecodeNullable<K, V>(
-  Map<K, V> enumValues,
-  dynamic source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    return null;
-  }
-  return _$enumDecode<K, V>(enumValues, source, unknownValue: unknownValue);
-}
 
 const _$StatusEnumMap = {
   Status.delivered: 'delivered',
