@@ -15,7 +15,7 @@ Room _$RoomFromJson(Map<String, dynamic> json) => Room(
           .toList(),
       metadata: json['metadata'] as Map<String, dynamic>?,
       name: json['name'] as String?,
-      type: _$enumDecode(_$RoomTypeEnumMap, json['type']),
+      type: $enumDecodeNullable(_$RoomTypeEnumMap, json['type']),
       updatedAt: json['updatedAt'] as int?,
       users: (json['users'] as List<dynamic>)
           .map((e) => User.fromJson(e as Map<String, dynamic>))
@@ -34,35 +34,8 @@ Map<String, dynamic> _$RoomToJson(Room instance) => <String, dynamic>{
       'users': instance.users.map((e) => e.toJson()).toList(),
     };
 
-K _$enumDecode<K, V>(
-  Map<K, V> enumValues,
-  Object? source, {
-  K? unknownValue,
-}) {
-  if (source == null) {
-    throw ArgumentError(
-      'A value must be provided. Supported values: '
-      '${enumValues.values.join(', ')}',
-    );
-  }
-
-  return enumValues.entries.singleWhere(
-    (e) => e.value == source,
-    orElse: () {
-      if (unknownValue == null) {
-        throw ArgumentError(
-          '`$source` is not one of the supported values: '
-          '${enumValues.values.join(', ')}',
-        );
-      }
-      return MapEntry(unknownValue, enumValues.values.first);
-    },
-  ).key;
-}
-
 const _$RoomTypeEnumMap = {
   RoomType.channel: 'channel',
   RoomType.direct: 'direct',
   RoomType.group: 'group',
-  RoomType.unsupported: 'unsupported',
 };
